refactor: remove legacy Item components and migrate to unified system
- Move RemoveFromRecentButton to ~/sections/common/components/buttons/
- Remove ExternalItemEditModal wrapper component
- Remove unused ItemContext.tsx
- Migrate RecipeEditModal to use UnifiedItemEditModal directly with proper conversions
- Remove entire ~/sections/food-item/ directory as all components migrated
- Update imports in TemplateSearchResults for new RemoveFromRecentButton location
- All Item/ItemGroup view components now use unified system exclusively
